+++ src/app.rs
@@ -526,7 +526,7 @@ impl App {
                 );
             }
 
-            self.did_config_fail_to_save = self.update_config_file().is_err();
+            // self.did_config_fail_to_save = self.update_config_file().is_err();
         }
     }
 
@@ -586,7 +586,7 @@ impl App {
                 );
             }
 
-            self.did_config_fail_to_save = self.update_config_file().is_err();
+            // self.did_config_fail_to_save = self.update_config_file().is_err();
         }
     }
 
@@ -642,7 +642,7 @@ impl App {
                 );
             }
 
-            self.did_config_fail_to_save = self.update_config_file().is_err();
+            // self.did_config_fail_to_save = self.update_config_file().is_err();
         }
     }
 
@@ -1565,28 +1565,28 @@ impl App {
         self.is_force_redraw = true;
     }
 
+    /// TODO: Disabled.
     /// Call this whenever the config value is updated!
-    fn update_config_file(&mut self) -> anyhow::Result<()> {
-        // TODO: Disabled.
-        // if self.app_config_fields.no_write {
-        //     // debug!("No write enabled.  Config will not be written.");
-        //     // Don't write!
-        //     // FIXME: [CONFIG] This should be made VERY clear to the user... make a thing saying "it will not write due to no_write option"
-        //     Ok(())
-        // } else if let Some(config_path) = &self.config_path {
-        //     // Update
-        //     // debug!("Updating config file - writing to: {:?}", config_path);
-        //     std::fs::File::create(config_path)?
-        //         .write_all(self.config.get_config_as_bytes()?.as_ref())?;
-        //     Ok(())
-        // } else {
-        //     // FIXME: [CONFIG] Put an actual error message?
-        //     Err(anyhow::anyhow!(
-        //         "Config path was missing, please try restarting bottom..."
-        //     ))
-        // }
-        Ok(())
-    }
+    // fn update_config_file(&mut self) -> anyhow::Result<()> {
+    //     if self.app_config_fields.no_write {
+    //         // debug!("No write enabled.  Config will not be written.");
+    //         // Don't write!
+    //         // FIXME: [CONFIG] This should be made VERY clear to the user... make a thing saying "it will not write due to no_write option"
+    //         Ok(())
+    //     } else if let Some(config_path) = &self.config_path {
+    //         // Update
+    //         // debug!("Updating config file - writing to: {:?}", config_path);
+    //         std::fs::File::create(config_path)?
+    //             .write_all(self.config.get_config_as_bytes()?.as_ref())?;
+    //         Ok(())
+    //     } else {
+    //         // FIXME: [CONFIG] Put an actual error message?
+    //         Err(anyhow::anyhow!(
+    //             "Config path was missing, please try restarting bottom..."
+    //         ))
+    //     }
+    //     Ok(())
+    // }
 
     pub fn kill_highlighted_process(&mut self) -> Result<()> {
         if let BottomWidgetType::Proc = self.current_widget.widget_type {